revising to use the version of edav_io in the current version of dev

+++ R/dal.R
@@ -99,7 +99,7 @@ edav_io <- function(
     }
 
     return(AzureStor::list_storage_files(azcontainer, file_loc) |>
-      dplyr::as_tibble())
+             dplyr::as_tibble())
   }
 
   if (io == "exists.dir") {
@@ -132,13 +132,34 @@ edav_io <- function(
     }
 
     if (grepl(".rds", file_loc)) {
+      corrupted.rds <- NULL
 
-      dest <- tempfile()
-      AzureStor::storage_download(container = azcontainer, file_loc, dest)
-      x <- readRDS(dest)
-      unlink(dest)
-      return(x)
+      tryCatch(
+        {
+          return(suppressWarnings(AzureStor::storage_load_rds(azcontainer, file_loc)))
+          corrupted.rds <<- FALSE
+        },
+        error = function(e) {
+          cli::cli_alert_warning("RDS download from EDAV was corrupted, downloading directly...")
+          corrupted.rds <<- TRUE
+        }
+      )
 
+      if (corrupted.rds) {
+        dest <- tempfile()
+        AzureStor::storage_download(container = azcontainer, file_loc, dest)
+        x <- readRDS(dest)
+        unlink(dest)
+        return(x)
+      }
+    }
+
+    if (grepl(".csv", file_loc)) {
+      return(suppressWarnings(AzureStor::storage_read_csv(azcontainer, file_loc)))
+    }
+
+    if (grepl(".rda", file_loc)) {
+      return(suppressWarnings(AzureStor::storage_load_rdata(azcontainer, file_loc)))
     }
   }
